Rename ConfigurationResponse to RemoteConfiguration; nest Manifest/ConfigItem
